# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SIEM/SOAR Platform API
  description: |
    Enterprise SIEM/SOAR Platform API providing security event management,
    detection, automated response, and AI-powered security operations.

    ## Features
    - Event ingestion and processing (100K+ EPS)
    - Real-time threat detection with YARA-L rules
    - Automated incident response with SOAR playbooks
    - AI-powered alert triage and investigation
    - Natural language query interface

    ## Authentication
    All API endpoints require authentication via API Key or JWT Bearer token.

  version: 1.0.0
  contact:
    name: Security Platform Team
    email: security-platform@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.siem.example.com/v1
    description: Production
  - url: https://api.staging.siem.example.com/v1
    description: Staging
  - url: http://localhost:8080/api/v1
    description: Local Development
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /rules:
    get:
      tags: [Rules]
      summary: List detection rules
      operationId: listRules
      parameters:
        - name: enabled
          in: query
          schema:
            type: boolean
        - name: severity
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of rules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleListResponse'

    post:
      tags: [Rules]
      summary: Create detection rule
      operationId: createRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleCreate'
      responses:
        '201':
          description: Rule created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Rule:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        severity:
          type: string
        query:
          type: string
        enabled:
          type: boolean
    RuleCreate:
      type: object
      required: [name, query, severity]
      properties:
        name:
          type: string
        description:
          type: string
        severity:
          type: string
        query:
          type: string
        enabled:
          type: boolean
          default: true
    RuleListResponse:
      type: object
      properties:
        rules:
          type: array
          items:
            $ref: '#/components/schemas/Rule'
...